- bug fixing
git-svn-id: svn://svn-srv/skosigns/trunk@148 702c9929-26ad-4ec3-b8b7-334534ed37ec

+++ src/org/openstreetmap/josm/plugins/scoutsigns/gui/details/filter/RoadSignFilterPanel.java
@@ -328,7 +328,7 @@ class RoadSignFilterPanel extends JPanel {
         pickerTo.setDate(null);
         pnlStatus.clearSelection();
         txtDupl.setText("");
-        txtConf.setText("" + SearchFilter.DEF_CONFIDENCE);
+        txtConf.setText("" + SearchFilter.DEF_CONFIDENCE.intValue());
         txtUsername.setText("");
         txtAppName.setText("");
         txtAppVers.setText("");
